feat(socket): fix socket wrapper
